pyproject.toml
[tool.poetry]
name = "linebot"
version = "0.1.0"
description = "LineBot Develope"
authors = ["Marcus_Lin <marcus.lin@e-elements.com.tw>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11"
flask = "3.0.3"
line-bot-sdk = "3.11.0"
pyngrok = "7.1.6"
requests = "2.31.0"
docx = "0.2.4"
langchain = "0.2.3"
openai = "1.33.0"
python-docx = "1.1.2"
pytelegrambotapi = "4.19.1"
langchain-community = "0.2.4"
chromadb = "0.5.0"
tiktoken = "0.7.0"
langchain-openai = "0.1.8"
langchain-cli = "0.0.24"
tenacity = "<8.4.0"
cassandra-driver = "3.29.1"
beautifulsoup4 = "4.12.3"
flask-wtf = "1.2.1"
flask-swagger-generator = "0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
